Added default import

+++ src/main/java/com/cursodevsuperior/dslist/Game.java
@@ -13,7 +13,7 @@ public class Game {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long id;
+    public Integer id;
     public String title;
     @Column(name = "game_year")
     public String year;
@@ -27,7 +27,7 @@ public class Game {
     public Game() {
     }
 
-    public Game(Long id, String title, String year, String genre, String platforms, String score, String imgUrl,
+    public Game(Integer id, String title, String year, String genre, String platforms, String score, String imgUrl,
             String shortDescription, String longDescription) {
         this.id = id;
         this.title = title;
@@ -40,11 +40,11 @@ public class Game {
         this.longDescription = longDescription;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
